refs #154 fixed centos build issues
git-svn-id: https://devel.codership.com/svn/galera/trunk@1796 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galera/src/mm_provider.hpp
@@ -196,7 +196,8 @@ namespace galera
             public:
                 size_t operator()(Transition const& tr) const
                 {
-                    return (gu::HashValue(tr.from_) ^ gu::HashValue(tr.to_));
+		  return (gu::HashValue(static_cast<int>(tr.from_))
+			  ^ gu::HashValue(static_cast<int>(tr.to_)));
                 }
             };
         private:
